requirements.txt
django==4.1
djangorestframework
djangorestframework-simplejwt
celery
flake8
rich
redis
python-dotenv
Pillow
drf-yasg
asgiref
django-channels
